requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
psycopg2==2.9.4
python-dotenv==0.21.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
zipp==3.9.0
